1.fix density sampling error in volumetric cloud 2.use cone tracing to calculate cube in PBR

+++ CarbonRender/Src/CRVolumetricCloudPass.cpp
@@ -7,7 +7,7 @@ void VolumetricCloudPass::GetReady4Render(PassOutput * input)
 
 	GLuint couldRt;
 	WindowSize size = WindowManager::Instance()->GetWindowSize();
-	couldRt = SetGLRenderTexture(size.w, size.h, GL_RGBA32F, GL_RGBA, GL_FLOAT, GL_COLOR_ATTACHMENT0);
+	couldRt = SetGLRenderTexture(size.w, size.h, GL_RGBA16F, GL_RGBA, GL_FLOAT, GL_COLOR_ATTACHMENT0);
 	dBuffer = SetGLDepthBuffer(size.w, size.h);
 
 	GLenum drawBuffers[1] = { GL_COLOR_ATTACHMENT0 };
